delete method

+++ src/frontend/components/deployment/DeployActionView.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { V1Deployment } from '@backend/k8s/model/V1Deployment'
+import { DialogHelper } from '@frontend/service/page/DialogHelper'
 import { Edit, RedoAlt, Trash } from '@vicons/fa'
 import { useDialog, useMessage } from 'naive-ui'
 import type { ActionMenuOption } from '@backend/model/actionMenu'
@@ -42,19 +43,11 @@ function getOptions(): ActionMenuOption[] {
       label: 'Delete',
       key: 'Delete',
       icon: Trash,
-      action: () => dialog.warning({
-        title: '警告',
-        content: `你确定删除：${props.deploy.metadata.namespace}/${props.deploy.metadata.name}吗?`,
-        positiveText: '删除',
-        negativeText: '放弃',
-        onPositiveClick: async () => {
+      action: () =>
+        DialogHelper.instance.dialog(dialog).confirmWithTarget('删除', `${props.deploy.metadata.namespace}/${props.deploy.metadata.name}`, async () => {
           await K8sService.deploymentService.deleteDeployments([`${props.deploy.metadata.namespace}/${props.deploy.metadata.name}`])
           drawer.close()
-        },
-        onNegativeClick: () => {
-          message.error('放弃删除')
-        },
-      }),
+        }),
     },
   ]
 }
