+++ src/frontend/components/node/NodeActionView.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import { DialogHelper } from '@frontend/service/page/DialogHelper'
 import { Edit, FireExtinguisher, PauseCircleRegular, PlayCircleRegular, Trash } from '@vicons/fa'
 import { useDialog, useMessage } from 'naive-ui'
 import type { ActionMenuOption } from '@backend/model/actionMenu'
@@ -75,21 +76,11 @@ function getOptions(): ActionMenuOption[] {
       label: 'Delete',
       key: 'Delete',
       icon: Trash,
-      action: () => {
-        dialog.warning({
-          title: '警告',
-          content: `你确定删除Node：${props.node.metadata.name}吗?`,
-          positiveText: '删除',
-          negativeText: '放弃',
-          onPositiveClick: async () => {
-            await K8sService.nodeService.deleteNode(props.node.metadata.name)
-            drawer.close()
-          },
-          onNegativeClick: () => {
-            message.error('放弃删除')
-          },
-        })
-      },
+      action: () =>
+        DialogHelper.instance.dialog(dialog).confirmWithTarget('删除', `Node：${props.node.metadata.name}`, async () => {
+          await K8sService.nodeService.deleteNode(props.node.metadata.name)
+          drawer.close()
+        }),
     },
   ]
 }
