+++ src/frontend/components/pod/PodActionView.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import { DialogHelper } from '@frontend/service/page/DialogHelper'
 import { Edit, StickyNoteRegular, Terminal, Trash } from '@vicons/fa'
 import { useDialog, useMessage } from 'naive-ui'
 import type { ActionMenuOption } from '@backend/model/actionMenu'
@@ -53,19 +54,11 @@ function getOptions(): ActionMenuOption[] {
       label: 'Delete',
       key: 'Delete',
       icon: Trash,
-      action: () => dialog.warning({
-        title: '警告',
-        content: `你确定删除Pod：${props.pod.metadata.namespace}/${props.pod.metadata.name}吗?`,
-        positiveText: '删除',
-        negativeText: '放弃',
-        onPositiveClick: async () => {
+      action: () =>
+        DialogHelper.instance.dialog(dialog).confirmWithTarget('删除', `Pod：${props.pod.metadata.namespace}/${props.pod.metadata.name}`, async () => {
           await K8sService.podService.deletePods([`${props.pod.metadata.namespace}/${props.pod.metadata.name}`])
           drawer.close()
-        },
-        onNegativeClick: () => {
-          message.error('放弃删除')
-        },
-      }),
+        }),
     },
   ]
 }
